+++ src/Helmholtz/helmkernels.f
@@ -74,7 +74,7 @@ cf2py intent(out) pot
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),pot(nd,nt)
@@ -85,7 +85,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d
       complex *16 zkeye,eye,ztmp
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -169,7 +169,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 charge(nd,ns),pot(nd,nt),grad(nd,3,nt)
@@ -181,7 +181,7 @@ c
       real *8 zdiff(3),dd,d
       complex *16 zkeye,eye,cd,cd1,ztmp
       complex *16 ztmp1,ztmp2,ztmp3
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -275,7 +275,7 @@ cf2py intent(out) pot
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -287,7 +287,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv
       complex *16 zkeye,eye,cd,cd1,dotprod
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -387,7 +387,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -399,7 +399,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv,dinv2
       complex *16 zkeye,eye,cd,cd2,cd3,cd4,dotprod
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -503,7 +503,7 @@ c
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -515,7 +515,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv
       complex *16 zkeye,eye,cd,cd1,dotprod
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
@@ -619,7 +619,7 @@ cf2py intent(out) pot,grad
 c
 cc      calling sequence variables
 c  
-      integer ns,nt,nd
+      integer *8 ns,nt,nd
       complex *16 zk
       real *8 sources(3,ns),ztarg(3,nt)
       complex *16 dipvec(nd,3,ns)
@@ -631,7 +631,7 @@ cc     temporary variables
 c
       real *8 zdiff(3),dd,d,dinv,dinv2
       complex *16 zkeye,eye,cd,cd2,cd3,cd4,dotprod
-      integer i,j,idim
+      integer *8 i,j,idim
       data eye/(0.0d0,1.0d0)/
 
       zkeye = zk*eye
